time conversion and time complexity

+++ hackerrank/mini-max-sum/minimaxsum.go
@@ -12,14 +12,14 @@ func MiniMaxSum(arr []int32) {
 	sort.Slice(arr, func(i, j int) bool {
 		return arr[i] < arr[j]
 	})
-	ff := int32(0)
-	lf := int32(0)
-	lens := int32(len(arr))
-	for i := int32(0); i < 4; i++ {
-		ff += arr[i]
+	ff := int64(0)
+	lf := int64(0)
+	lens := int64(len(arr))
+	for i := int64(0); i < 4; i++ {
+		ff += int64(arr[i])
 		if i == 1 {
-			for j := int32(1); j < lens; j++ {
-				lf += arr[j]
+			for j := int64(1); j < lens; j++ {
+				lf += int64(arr[j])
 			}
 		}
 	}
